+++ modules/parallel/tests/test_fetchOutputs.m
@@ -7,8 +7,6 @@
 % SPDX-License-Identifier: LGPL-3.0-or-later
 % LICENCE_BLOCK_END
 %=============================================================================
-% <--SEQUENTIAL TEST REQUIRED-->
-%=============================================================================
 p = str2func('cos');
 b = backgroundPool();
 A = parfeval(b, p, 1, pi);