+++ CHANGELOG.md
@@ -34,6 +34,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - matrix of handles supported.
 
+- [#676](http://github.com/Nelson-numerical-software/nelson/issues/676): use fmtlib 9.0
+
 ### Fixed
 
 - [#659](http://github.com/Nelson-numerical-software/nelson/issues/659): qhelpgenerator on macOs after brew Qt6.3.0 update.